updates to the results table
