+++ src/App.jsx
@@ -11,8 +11,8 @@ import {
 import Main from "./components/Main/Main";
 import SplashScreenView from "./views/SplashScreenView/SplashScreenView";
 import ArchitectureView from "./views/ArchitectureView/ArchitectureView";
-import ResultsView from "./views/ResultsView/ResultsView";
-import RunTestsView from "./views/RunTestsView/RunTestsView";
+import ResultsView from "./views/FileRebuildPerformanceTest/ResultsView/ResultsView";
+import RunTestsView from "./views/FileRebuildPerformanceTest/RunTestsView/RunTestsView";
 
 import styles from "./App.module.scss";
