+++ src/views/FileRebuildPerformanceTest/ResultsView/ResultsTable/ResultsTable.module.scss
@@ -1,0 +1,9 @@
+.resultRow {
+    padding-bottom: 3rem;
+
+    .timestamp {
+        font-size: 2rem;
+        font-weight: 1rem;
+        padding-bottom: 1rem;
+    }
+}
